Misc minor improvements
Added an FX playing for successful uses of the parrying skill
Successfully parrying an attack will send a sysmessage to the player informing them of their feat, if DISPLAYHITMSG is enabled in UOX.INI
Updated dictionary files with new entry about blocking attacks using parrying skill
Fixed placing a house addon teleporting player to the location of the addon
Added support for TELEPORT tag to regions.dfn. If set to 0, use of the Teleport spell is not allowed in a given region. Defaults to 1 if not specified.
Exposed canTeleport JS property for regions
Updated regions.dfn to disable usage of Teleport spell within level 3 of the Wisp dungeon
Updated regions.dfn to enable usage of RECALL spell in city/dungeon of Wind (so players can recall out of there)
Added missing UOX.INI settings for feature negotation, as mentioned in previous commit!