Crafting OverHual
I updated all blacksmith, Carpentry, Tailoring, Alchemy, fletching. to use new Menus that is provided in Pub15 right before AOS release.

In CraftTool you can turn off new menus and just use older UOX3 Menus.

I also Updated the Create DFN scripts so they are correct with new Pub 15 updates as well.

This has a lot of functions and a lot of math was done to get the formulas correct in the itemdetailgump.
This system can be updated very easily and adding a new craft should be simple with a small user guide I will create on the forums. Explaining how to add custom crafts to this system.