Update Magic dfn files
I added each one  with a name of the item, you can do either id or name in game to add the item now.

Also fixed couple scrolls that had weight=1, that wasn't needed because base scroll already had weight=1

fixed a scroll that had a space after the } bracket.